gen: zero init
